Retrieve managed configurations in the Splash activity
