+++ owncloudApp/src/main/java/com/owncloud/android/ui/activity/FileDisplayActivity.kt
@@ -36,7 +36,6 @@ import android.content.ComponentName
 import android.content.Context
 import android.content.Intent
 import android.content.IntentFilter
-import android.content.RestrictionsManager
 import android.content.ServiceConnection
 import android.content.res.Resources.NotFoundException
 import android.net.Uri
@@ -52,7 +51,6 @@ import com.owncloud.android.AppRater
 import com.owncloud.android.BuildConfig
 import com.owncloud.android.MainApp
 import com.owncloud.android.R
-import com.owncloud.android.data.preferences.datasources.implementation.SharedPreferencesProviderImpl
 import com.owncloud.android.databinding.ActivityMainBinding
 import com.owncloud.android.datamodel.FileDataStorageManager
 import com.owncloud.android.datamodel.OCFile
@@ -94,8 +92,6 @@ import com.owncloud.android.ui.preview.PreviewImageFragment
 import com.owncloud.android.ui.preview.PreviewTextFragment
 import com.owncloud.android.ui.preview.PreviewVideoActivity
 import com.owncloud.android.ui.preview.PreviewVideoFragment
-import com.owncloud.android.utils.CONFIGURATION_SERVER_URL
-import com.owncloud.android.utils.CONFIGURATION_SERVER_URL_INPUT_VISIBILITY
 import com.owncloud.android.utils.Extras
 import com.owncloud.android.utils.PreferenceUtils
 import kotlinx.coroutines.CoroutineScope
@@ -152,10 +148,6 @@ class FileDisplayActivity : FileActivity(), FileFragment.ContainerActivity, OnEn
         Timber.v("onCreate() start")
 
         super.onCreate(savedInstanceState) // this calls onAccountChanged() when ownCloud Account is valid
-
-        if (BuildConfig.FLAVOR == MainApp.MDM_FLAVOR) {
-            handleRestrictions()
-        }
 
         checkPasscodeEnforced(this)
 
@@ -223,24 +215,6 @@ class FileDisplayActivity : FileActivity(), FileFragment.ContainerActivity, OnEn
 
         if (resources.getBoolean(R.bool.enable_rate_me_feature) && !BuildConfig.DEBUG) {
             AppRater.appLaunched(this, packageName)
-        }
-    }
-
-    private fun handleRestrictions() {
-        val restrictionsManager = getSystemService(Context.RESTRICTIONS_SERVICE) as RestrictionsManager
-        val restrictions = restrictionsManager.applicationRestrictions
-        cacheRestrictions(restrictions)
-    }
-
-    private fun cacheRestrictions(restrictions: Bundle) {
-        val preferencesProvider = SharedPreferencesProviderImpl(this)
-        if (restrictions.containsKey(CONFIGURATION_SERVER_URL)) {
-            val confServerUrl = restrictions.getString(CONFIGURATION_SERVER_URL)
-            confServerUrl?.let { preferencesProvider.putString(CONFIGURATION_SERVER_URL, it) }
-        }
-        if (restrictions.containsKey(CONFIGURATION_SERVER_URL_INPUT_VISIBILITY)) {
-            val confServerUrlInputVisibility = restrictions.getBoolean(CONFIGURATION_SERVER_URL_INPUT_VISIBILITY)
-            preferencesProvider.putBoolean(CONFIGURATION_SERVER_URL_INPUT_VISIBILITY, confServerUrlInputVisibility)
         }
     }
